project1/integrated search/requirements.txt
# requirements.txt

flask
requests
pandas
openpyxl
slack_sdk
python-dotenv
deep-translator
beautifulsoup4
feedparser
summa
Flask-Session